+++ tests/TestCase.php
@@ -9,22 +9,11 @@ use LaravelAIEvaluation\LaravelAIEvaluationServiceProvider;
 
 class TestCase extends \Orchestra\Testbench\TestCase
 {
-    protected function setUp(): void
-    {
-        parent::setUp();
-        // additional setup
-    }
-
     protected function getPackageProviders($app): array
     {
         return [
             AiServiceProvider::class,
             LaravelAIEvaluationServiceProvider::class,
         ];
-    }
-
-    protected function getEnvironmentSetUp($app)
-    {
-        // perform environment setup
     }
 }